Chatbots en WhatsApp API (MedPets)/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
httpx
openai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib